ft_striteri.c

+++ libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: yohwang <yohwang@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/02 17:04:02 by yohwang           #+#    #+#             */
-/*   Updated: 2021/12/02 19:40:07 by yohwang          ###   ########.fr       */
+/*   Updated: 2021/12/02 19:43:48 by yohwang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	int	i;
+	unsigned int	i;
 
 	if (s && f)
 	{
 		i = -1;
-		while ((unsigned int)++i < ft_strlen(s))
-			f((unsigned int)i, &s[i]);
+		while (++i < ft_strlen(s))
+			f(i, &s[i]);
 	}
 }
